ctype-compatibility.py: Make get_lines_from_file() work even if the files ends with a continuation line
Alexandre Oliva> get_lines_from_file doesn't check that current_line
Alexandre Oliva> is empty at the end of the loop.  that's fine,
Alexandre Oliva> because it would indicate a file terminated with a
Alexandre Oliva> continuation line.  I'm not sure this is worth
Alexandre Oliva> checking and reporting an error, but maybe a comment
Alexandre Oliva> would be in order