ctype-compatibility.py: make conversion from ctype_dict to ctype_dict2 more efficient
Alexandre Oliva> the conversion of ctype_dict to ctype_dict2 tests
Alexandre Oliva> whether key is in a list; I *think* testing the type
Alexandre Oliva> of value might be more efficient, while avoiding
Alexandre Oliva> hard-coding the list of maps.  Also, pulling this
Alexandre Oliva> test out of the loop can avoid repeating the test for
Alexandre Oliva> every loop iteration, even if duplicating the inner
Alexandre Oliva> loop in the if and else cases.

Readability suffers a bit but it is faster, 5.4 seconds
instead of 6.2 seconds for the whole script on my machine.