utf8-compatibility.py: Change detailed report on WIDTH to be more in the style as the new detailed report on CHARMAP
Alexandre Oliva> in check_width, it seems to me that the tests for
Alexandre Oliva> missing and added characters are redundant with the
Alexandre Oliva> change tests, since missing entries have width 1.

Yes, that was redundant. Pravin originally had only the tests
for missing and added and I found it hard to see when a
character was removed with a certain width (and therefore appeared in
“removed”) and was added again with a different width (and therefore
appeared in “added” at a line in the report far away from the line
where it was removed). So it was hard to see when a character changed
width from 1 to 2 or vice versa because it was shown by 2 lines
in the report far away from each other.

So I added the report of the “changed” characters
which actually listed all changes: It listed removing a character
as a change …->1 (because characters not in the file have the default
width 1) and adding a character as a change 1->… .
As this contained all changes, I thought of removing the report
for the added and removed characters, but Pravin liked these
so I left them in.

Now I have changed it again to behave like the new detailed
report for CHARMAP, now “removed”, “added”, and “changed”
are mutually exclusive:

    - removed lists characters which had width 0 or 2 and
      now have 1 because they are not in the file anymore
    - changed lists characters which changed from width 0 to width 2
      or vice versa (i.e. they were in the old file and are in the
      new file but the width has changed)
    - added lists characters which had width 1 because they were
      not in the file but have now with 0 or 2 because they
      have been added.

This seems actually easiest to read.